+++ CHANGELOG.md
@@ -491,3 +491,4 @@ All notable changes to Project Genesis will be documented in this file.
 
 - Implemented `is_lava_dancer` trait. Entities with this trait gain energy when on `lava` terrain.
 - Implemented `is_wall_dancer` trait. Entities with this trait gain energy when on `wall` terrain.
+- Implemented `is_wind_dancer` trait. Entities with this trait gain energy when inside a `storm` event.